MISRA: check rule [misra-c2012-10.6]
The value of a composite expression shall not be assigned to an object
with wider essential type

+++ src/operations.c
@@ -387,7 +387,7 @@ static inline bool parse_byte(uint8_t byte,
 
             // This is the one allowable non-reveal operation per set
 
-            out->operation.tag = (uint8_t) state->tag;
+            out->operation.tag = state->tag;
 
             // Deliberate epsilon-transition.
             state->op_step = STEP_OP_TYPE_DISPATCH;
